requirements.txt
APScheduler==3.7.0
asgiref==3.3.4
Django==3.2
django-cors-headers==3.7.0
django-redis==4.12.1
djangorestframework==3.12.4
numpy==1.19.5
pandas==1.1.5
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
schedule==1.1.0
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
tzlocal==2.1